+++ Example/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.2
+
+* Bootpay.confirm(data) -> Bootpay.transacionConfirm() 으로 수정 
+
+* confirm에서 return false 일 때 결제창 닫히는 버그 수정   
+
 ## 4.0.1
 
 * bootpay bio 호환을 위한 필드 scope 변경